tests/requirements-dev.txt
pytest
pytest-asyncio
python-dotenv
google-cloud-secret-manager
google-auth
tqdm
vertexai